requirements.txt
streamlit
easyocr
numpy
opencv-python-headless
pillow
torch==2.0.1
torchvision==0.15.2
